requirements.txt
torch
torchvision
tqdm
opencv-python
wandb
numpy
matplotlib
torchinfo
streamlit
kornia
pillow